[Refactor](schema-form): move formatError to helpers
