+++ packages/canard/schema-form/src/helpers/error/index.ts
@@ -1,1 +1,2 @@
 export { transformErrors } from './transformErrors';
+export { formatError } from './formatError';
